fix: :construction: Add helper to run networkx tests with old signature
I think we should actually update the read_nx and write_nx signatures? But in the meantime, I wrote a helper function to convert the lists of names, units, etc. into a list of Axes for constructing GeffMetadata. The tests still fail because of the user warning about units not being caught.