feat(read): Support rustworkx and spatial-graph backends in unified read function (#273)
# Proposed Change

- Add `RUSTWORKX` and `SPATIAL_GRAPH` to `SupportedBackend` enum.
- Update `construct_rx` to take array arguments not a single dict of
arrays.
- Ensure `construct_rx` and `construct_sg` are only imported into
`read.py` if the relevant backends have been installed. If someone tries
to use the `read` function with a backend that hasn't been installed an
error will be raised.
- Small update to `read` function signature: `backend_kwargs` are now
real python kwargs not a dict. This allows the overloads for each
backend to include the additional kwargs, e.g. spatial-graph overload
has the `position_attr` argument.
- Updated all the test helper functions for `test_read` for rustworkx
and spatial graph. Now all backends are tested identically.
- For spatial graph there is an additional `sg_get_node_spatial_props`
function because all the spatial props are combined into one position
prop, whereas the other backends (nx and rx) keep the different spatial
props (e.g. `"x"`, `"y"`, `"t"`) separate.

N.b. currently the whole of `test_io.py` will be skipped if any of the
backends are not installed because of the `pytests.importskip` at the
top of the file. There may be a better way to do this but as long as one
of the CIs runs with everything installed I think it should be fine.

Resolves #260.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- Read implementation

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

## If you changed the specification
- [ ] I have checked that any validation functions and tests reflect the
changes.
- [ ] I have updated the GeffMetadata and the json schema using `pytest
--update-schema` if necessary.
- [ ] I have updated docs/specification.md to reflect the change.
- [ ] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

## If you have added or changed an implementation
- [x] I wrote tests for the new implementation using standard fixtures
supplied in conftest.py.
- [x] I updated pyproject.toml with new dependencies if needed.
- [x] I added a function to tests/bench.py to benchmark the new
implementation.

---------

Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>
Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>
Co-authored-by: melisande-c <melisande.croft@fht.org>
Co-authored-by: Talley Lambert <talley.lambert@gmail.com>