API reorganization (#292)
# Proposed Change
Majorly reorganizes where code lives and what the import patterns are
with the goal of making our public facing api docs easy to navigate and
read. Closes #183. This will introduce breaking changes for libraries
that are depending on geff (sorry!)

Still to do, but will happen in another PR:
- A comprehensive review of how the docstrings are rendering and adding
some missing docstrings
- Reorganize tests to match the new module structure

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Documentation update
- Maintenance (e.g. dependencies, CI, releases, etc.)

Which topics does your change affect? Delete those that do not apply.

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

---------

Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>