Add a testing utility function for comparing two geffs on disk (#137)
# Proposed Change
This function checks that the types and shapes of the zarr arrays
contained within a geff match between two different geffs.

Also updates the networkx read/write consistency test to use this
function. When we update the rest of the networkx tests to use fixtures,
we should use this function there as well.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Tests

Which topics does your change affect? Delete those that do not apply.

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the [developer/contributing](../CONTRIBUTING) docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly.
- [x] If I changed the specification, I have checked that any validation
functions and tests reflect the changes.

---------

Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>
Co-authored-by: Georgeos Hardo <georgeoshardo@protonmail.com>