Reorganize tests to match new API organization (#304)
# Proposed Change
Closes #298 

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Tests

Which topics does your change affect? Delete those that do not apply.

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [ ] I have checked that I maintained or improved code coverage.