Improve framework for validation (#204)
# Proposed Change
This PR works towards the validation framework described here:
https://github.com/live-image-tracking-tools/geff/issues/152#issuecomment-3098634966.
It will ultimately incorporate some of the validation functions
contributed in #180.

New checks that have been implemented so far
- Check node id dtype, starting on #176 (still need to do the uint part
in another PR)
- Checks that properties specified in axes are actually present in the
zarr without missing data

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly.

# Further Comments

---------

Co-authored-by: Benjamin Gallusser <bgallusser@googlemail.com>
Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>