feat: ✨ add zarr version compatibility warnings (#313)
# Proposed Change
- Add zarr v2/v3 version detection and warnings in core_io utils
- Warn when using zarr v3 format with zarr v2 library
- Add comprehensive tests for zarr version warning functionality
- Minor: fix import references and update gitignore

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement
- Tests

Which topics does your change affect? Delete those that do not apply.
- Core io

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).