Document loading subsets of geff files (#317)
Closes #227 by providing examples of how to load either some properties
of a geff or either a subset of nodes/edges. The imports for the
construct functions are ugly at the moment and will change with #308. I
think we should expose a universal construct function along with read
and write that we expose at the top level geff import. But that can
happen in another PR.

~~Alternatively after #308 is merged I think we could update the code
blocks to something like this~~ Code blocks are now updated and ready to
merge in after #308 is merged.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Documentation update

Which topics does your change affect? Delete those that do not apply.

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.