Improve schema/metadata documentation (#324)
# Proposed Change
This PR improves how we document geff schema/metadata in the
specification. I removed the embedded html version of the schema and am
replacing it with direct documentation of the pydantic models. Some
massaging and refinement is needed, but I'd like to get feedback before
I go any further.

One open question is where certain types of information should live. For
example, for the axes metadata, there is a class docstring, the field
description in GeffMetadata and a block of text in the specification
docs. Currently I moved the block of text from the specification into
the pydantic model docstrings.

This PR should ultimately close #201 

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Documentation update

Which topics does your change affect? Delete those that do not apply.
- Specification

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

---------

Co-authored-by: Talley Lambert <talley.lambert@gmail.com>
Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>