Add dataframe interops (#185)
# Proposed Change
Add API to convert nodes and edges to pandas dataframe

# Types of Changes
- New feature or enhancement


# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [X] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly.

---------

Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>