Remove Affine and replace with Axis scale (#331)
# Proposed Change
Closes #274, #276 and #327. Following discussions in the 9/19/25
community meeting, we decided to remove the `Affine` metadata in favor
of simple scale and offset metadata. In offline discussion, @cmalinmayor
and I are proposing that we simply add two fields to the `Axis`
metadata: scale and scaled_unit.

Currently the default value for `scale` is None so that it is clear when
the user has not set the field themselves. We expect that downstream
tools would replace None with 1 if they are trying to apply the scale to
data. We are not providing utilities for applying the scale since it is
straightforward.

We decided that offset was not necessary to include since the use case
isn't as obvious with a graph as it might be with segmentation
annotations of an image. We can always add offset in the future if there
is a clear use case, but currently it would only add overhead related to
specifying whether the offset should be applied before or after changing
the scale.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement

Which topics does your change affect? Delete those that do not apply.
- Specification
- Metadata

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

## If you changed the specification
- [x] I have checked that any validation functions and tests reflect the
changes.
- [x] I have updated the GeffMetadata and the json schema using `pytest
--update-schema` if necessary.
- [x] I have updated docs/specification.md to reflect the change.
- [x] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

---------

Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>