Implement vlen_props to support variable-length content (#251)
# Proposed Change
This PR implements support for reading and writing variable-length
properties such as polygons. The variable length properties must be
numpy arrays of the same number of dimensions and dtype, but can have
any shape. There is a helper, `construct_var_len_props` that can convert
a sequence of ArrayLike or None into this format.
Variable length properties are stored with an additional `data` entry
that contains the data serialized into one long one-dimensional array of
the shared dtype. The `values` array contains the offset and shape of
the array for the corresponding node/edge. The `missing` array remains
unchanged, and there are no restrictions on what is in the `values`
array when the property is missing.

Strings are NOT encoded with this serialization scheme, but instead with
the zarr-python implementation of the zarr-extensions specification for
variable length strings.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- New feature or enhancement
- Documentation update
- Tests

Which topics does your change affect? Delete those that do not apply.
- Specification

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

## If you changed the specification
- [x] I have checked that any validation functions and tests reflect the
changes.
- [ ] I have updated the GeffMetadata and the json schema using `pixi
run update-json` if necessary.
- [x] I have updated docs/specification.md to reflect the change.
- [x] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

## If you have added or changed an implementation
- [x] I wrote tests for the new implementation using standard fixtures
supplied in conftest.py.
- [ ] I updated pyproject.toml with new dependencies if needed.
- [ ] I added a function to tests/bench.py to benchmark the new
implementation.

# Further Comments
Updated diagram coming soon

---------

Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>
Co-authored-by: Morgan Schwartz <msschwartz21@gmail.com>
Co-authored-by: Caroline Malin-Mayor <cmalinmayor@gmail.com>