Monorepo for separate versioning of geff specification and geff python implementation (#329)
# Proposed Change
Closes #164 and picks up from where Talley started in #326. A couple
things that still need to be taken care of:

- [x] Check that CI runs all tests in both packages
- [x] Figure out how the build/deploy to pypi action is going to work
and if we may need to deploy geff-spec manually
- [ ] On a longer term basis, we need to put some more thought into how
we are going to maintain different versions of the schema for backwards
compatibility both in the code base and in the docs, but I think we can
address that in the future when we get to the point of releasing new
schemas.
- [x] Question: should the json schema move into the geff-spec package
folder? -> I'm going to leave it where it is.
- [x] Question: how is the version of the specification reflected in the
geff-schema.json? -> If someone wants to find a geff-schema.json for a
specific version they can use the release tag on the repo, but the tag
won't be directly written in the schema file.

This PR has the potential to cause some merge headaches so I will try to
wait until as many open PRs are merged in as possible and just keep this
one up to date as new changes come in.

The release strategy following this PR will be a bit different and it
may take a try or two of actual releases to get the bugs ironed out.
- `geff` is still versioned based on the git tag and releases are
triggered by pushing a tag to github.
- `geff-spec` is versioned manually in the pyproject.toml. Releasing a
new version is a two part process. First update the pyproject.toml
version, then push the corresponding tag prefixed with `spec` to github.
- Release notes will be generated by a new action that should only
include PRs that are relevant to the specific package that is being
released. This is the change that is most likely to fail and require
some trial and error to get right, but it's pretty low risk.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Maintenance (e.g. dependencies, CI, releases, etc.)

Which topics does your change affect? Delete those that do not apply.
- Specification

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [ ] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [ ] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

## If you changed the specification
- [ ] I have checked that any validation functions and tests reflect the
changes.
- [ ] I have updated the GeffMetadata and the json schema using `pytest
--update-schema` if necessary.
- [ ] I have updated docs/specification.md to reflect the change.
- [ ] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

## If you have added or changed an implementation
- [ ] I wrote tests for the new implementation using standard fixtures
supplied in conftest.py.
- [ ] I updated pyproject.toml with new dependencies if needed.
- [ ] I added a function to tests/bench.py to benchmark the new
implementation.

# Further Comments
If this is a relatively large or complex change, kick off the discussion
by explaining why you chose the solution you did and what alternatives
you considered, etc...

---------

Co-authored-by: Talley Lambert <talley.lambert@gmail.com>