Expose axis scale and scaled_units in public write API (#344)
# Proposed Change
Closes #343 

I made a few changes that weren't directly related. Previously we
performed some handling and validation of axis names in the
`write_dicts` function. I deleted that so that all axis metadata
handling is performed in the `Backend.write` function for consistency.
This change also necessitated a change to the function where we compute
axis min/max to do a check that the axis actually exists in the node
props.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Bugfix (non-breaking change which fixes an issue)

Which topics does your change affect? Delete those that do not apply.
- Implementations (`geff._graph_libs`)

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).