Validate when writing geffs (#342)
# Proposed Change
Closes #333

I considered a few different approaches to how to make sure we didn't
write an invalid geff. A major constraint is that our structural
validation function takes a StoreLike as input which means that the geff
has to be written before we can validate the structure. Given this
constraint, I introduced a utility function that deletes a geff that we
have written while doing its best to handle the different possible types
in StoreLike and avoiding delete any members of the zarr that aren't
geff members.

Alternatively, I considered generalizing the `validate_structure`
function to operate on either a StoreLike or an InMemoryGeff but the
approaches that I tried were leading to a ton of code duplication and
maintenance overhead. I think the deleting option is going to be easier
to maintain in the long run even if its slightly less elegant and a bit
inefficient.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Bugfix (non-breaking change which fixes an issue)

Which topics does your change affect? Delete those that do not apply.
- Core io
- Validate

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

---------

Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>