Eliminate overwrite behavior and instead check for existing geff before writing (#364)
# Proposed Change
Closes #361. If overwrite is True, deletes the geff and writes a new
geff. Exposed in `geff.write`, all convert functions and `write_arrays`.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Bugfix (non-breaking change which fixes an issue)

Which topics does your change affect? Delete those that do not apply.
- Convert
- Core io

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

---------

Co-authored-by: Caroline Malin-Mayor <malinmayorc@janelia.hhmi.org>