Docs: Foreign type links & other fixes (#375)
# Proposed Change

I was only going to fix a couple small things but then I saw more things
to fix 😁

- I removed an outdated example from `GeffReader` and instead put a link
to the *Tips and Tricks* page which has examples on how to use it.
- I added all the foreign type object inventories to the mkdocs plugin
mkdocstrings.
- I added very basic one line docstrings to the `geff-spec` type aliases
`SpaceUnits`, `TimeUnits`, `AxisType` and `DTypes` so that they are
rendered by mkdocs and can be linked to.
- Similarly I added a docstrings to `SupportedGraphs` and `InMemoryGeff`
and imported them into the top level api.
- I fixed lots of docstring types so that they can be automatically
linked to external and internal types.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Documentation update

Which topics does your change affect? Delete those that do not apply.
- Docs

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [ ] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [ ] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).

## If you changed the specification
- [ ] I have checked that any validation functions and tests reflect the
changes.
- [ ] I have updated the GeffMetadata and the json schema using `pytest
--update-schema` if necessary.
- [ ] I have updated docs/specification.md to reflect the change.
- [ ] I have updated implementations to reflect the change. (This can
happen in separate PRs on a feature branch, but must be complete before
merging into main.)

## If you have added or changed an implementation
- [ ] I wrote tests for the new implementation using standard fixtures
supplied in conftest.py.
- [ ] I updated pyproject.toml with new dependencies if needed.
- [ ] I added a function to tests/bench.py to benchmark the new
implementation.