Dask future warning (#389)
# Proposed Change
Tests with latest released version of dask were failing because of a
future warning in the to_zarr function signature.

Options:
  1. Ignore the warning (what I chose for now)
2. Update to the new signature and pin to >= the release from last week
(seemed too prohibitive)
3. Do a dask version check the same way we do a zarr version check and
pass the arguments based on version

Depending on when dask actually removes the backward compatible **kwargs
pattern, we may still need to do option 3 instead of option 2, but I'd
rather wait and see if it is necessary. To detect the actual breaking
changed, I also added weekly scheduled runs of the testing CI.

# Types of Changes
What types of changes does your code introduce? Delete those that do not
apply.
- Maintenance (e.g. dependencies, CI, releases, etc.)

Which topics does your change affect? Delete those that do not apply.
- Convert (where we are using dask.to_zarr)
- Testing

# Checklist
Put an x in the boxes that apply. You can also fill these out after
creating the PR. If you're unsure about any of them, don't hesitate to
ask. We're here to help! This is simply a reminder of what we are going
to look for before merging your code.

- [x] I have read the
[developer/contributing](https://github.com/live-image-tracking-tools/geff/blob/main/CONTRIBUTING)
docs.
- [x] I have added tests that prove that my feature works in various
situations or tests the bugfix (if appropriate).
- [x] I have checked that I maintained or improved code coverage.
- [x] I have written docstrings and checked that they render correctly
by looking at the docs preview (link left as a comment on the PR).